backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.2.0
Pillow==9.5.0
torch==2.0.0
torchvision==0.15.0
opencv-python==4.7.0.72
numpy==1.24.0
python-dotenv==1.0.0
Pillow==9.5.0
django-cleanup==7.0.0